requirements/local.txt
# includes the _base.txt requirements file
-r requirements.txt
coverage
django-debug-toolbar
django-shell-plus
grin
pyflakes
pynliner
python-ptrace
Werkzeug
